// website/skeet-doc/src/types/article.ts
export type BlogIndex = {
  title: string
  category: string
  thumbnail: string
  date: string
}
export type BlogContent = {
  title: string
  asPath: string
  category: string
  thumbnail: string
  date: string
  content: string
}

export type NewsIndex = {
  title: string
  category: string
  thumbnail: string
  date: string
  content: string
}
export type NewsContent = {
  title: string
  asPath: string
  category: string
  thumbnail: string
  date: string
  content: string
}

export type DocIndex = {
  title: string
  description: string
}

export type DocContent = {
  title: string
  description: string
  content: string
}

export type LegalContent = {
  title: string
  asPath: string
  description: string
  content: string
}